Check for SentDate before moving email to sending (see description)
Its important to create a sentDate first before anything else. In a rare case where an email without sentDate get moved to the sending box but something goes wrong before a sentDate gets created, then the email will remain lost in a void where it remains in that state since the email-service cannot check if a timeout has been reached to recycle the mail. As it does not have a sentDate but sits in the sending box the email service will just ignore that email and it will never actually get send or moved to the failbox